pyproject.toml
[tool.poetry]
name = "junglescout-scrapper"
version = "0.1.0"
description = ""
authors = ["federicober <federico.ober@hotmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
rich = "^11.2.0"
beautifulsoup4 = "^4.10.0"
typer = "^0.4.1"
selenium = "^4.3.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
black = "isort"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
